+++ src/airfoil_aero/AirfoilAero.jl
@@ -8,7 +8,8 @@ using NPZ
 using Xfoil
 using Printf: @sprintf
 using ..VortexStepMethod: SectionAero, interpolate_matrix_nans!, delta_suffix,
-                          write_node_rows, section_surface, set_taylor_polar!
+                          write_node_rows, section_surface, set_taylor_polar!,
+                          KulfanParameters
 
 include("kulfan.jl")
 include("deform.jl")